+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 * `Checkpoint-Migration` will no longer delete migration scripts that have been pushed to the database. Instead it
   will now export rows from the `rivet.Migrations` table and include them in the `schema.ps1` file.
-* Added an `Initialize-Schema` switch to the `Invoke-Rivet` script that is used to initialize database(s) with the
+* Added an `InitializeSchema` switch to the `Invoke-Rivet` script that is used to initialize database(s) with the
   `schema.ps1` file that is generated from `Checkpoint-Migration`.